requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyYAML==6.0
redis==4.4.2
requests==2.28.2
urllib3==1.26.14
Werkzeug==2.2.2
